refactor(admin): move search endpoints under /admin/search
