+++ apps/core/src/routes/v1/admin/index.ts
@@ -4,6 +4,7 @@ import { OpenAPIHonoWithAuth } from "@/lib/hono";
 import { requireAdminAuthContext } from "@/middleware/auth";
 import invoicesRouter from "./invoices/index.js";
 import organizationsRouter from "./organizations/index.js";
+import searchRouter from "./search/index.js";
 import tasksRouter from "./tasks/index.js";
 import usersRouter from "./users/index.js";
 
@@ -15,6 +16,7 @@ const requireAdmin = createMiddleware(async (c, next) => {
 const app = new OpenAPIHonoWithAuth();
 
 app.use("*", requireAdmin);
+app.route("/search", searchRouter);
 app.route("/users", usersRouter);
 app.route("/organizations", organizationsRouter);
 app.route("/invoices", invoicesRouter);
