+++ apps/core/src/routes/v1/admin/search/index.ts
@@ -1,0 +1,11 @@
+import { OpenAPIHonoWithAuth } from "@/lib/hono";
+
+import mountSearchAdminOrganizations from "./organizations/get.js";
+import mountSearchAdminUsers from "./users/get.js";
+
+const app = new OpenAPIHonoWithAuth();
+
+mountSearchAdminUsers(app);
+mountSearchAdminOrganizations(app);
+
+export default app;
