+++ apps/core/src/routes/v1/admin/users/index.ts
@@ -1,11 +1,9 @@
 import { OpenAPIHonoWithAuth } from "@/lib/hono";
 
 import mountListAdminUsers from "./get.js";
-import mountSearchAdminUsers from "./search/get.js";
 
 const app = new OpenAPIHonoWithAuth();
 
 mountListAdminUsers(app);
-mountSearchAdminUsers(app);
 
 export default app;
